fix datatypes for very large domains

+++ src/databases/IDX/pidx_idx_io.cpp
@@ -336,10 +336,10 @@ unsigned char* PIDXIO::getData(const VisitIDXIO::Box box, const int timestate, c
   int v_per_sample = 0;
   PIDX_values_per_datatype(variable->type_name, &v_per_sample, &bits_per_sample);
 
-  uint32_t this_size = (uint32_t)(local_size[0] * local_size[1] * local_size[2]);
+  size_t this_size = (size_t)(local_size[0] * local_size[1] * local_size[2]);
 
-  void *data = malloc((uint32_t)((bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
-  memset(data, 0, ((uint32_t)(bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
+  void *data = malloc((size_t)((bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
+  memset(data, 0, ((size_t)(bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
 
   ret = PIDX_variable_read_data_layout(variable, local_offset, local_size, data, PIDX_row_major);
   if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_read_data_layout");
